Issue #4437: Modified IllegalImportCheckTest.java and moved its input files to the illegalimport subdirectory
